requirements.txt
opencv-python
numpy
Pillow
onnxruntime ; python_version>="3.8"
